+++ actionview/CHANGELOG.md
@@ -1,3 +1,15 @@
+*   Add `reset_tag` helper method to reset the form tag state.
+
+    This method is useful when you want to reset the form tag state to the default values.
+
+    ```erb
+    <%= reset_tag %>
+    <%= reset_tag('Clear Form') %>
+    <%= reset_tag('Reset', class: 'btn btn-secondary') %>
+    ```
+
+    *Akhil G Krishnan*
+
 *   Rename `text_area` methods into `textarea`
 
     Old names are still available as aliases.